+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Prevent to move the default item template to a subfolder
   (removal was already managed, now moval is not possible neither).
   [gbastien]
+- Display a help message on the item view regarding copy groups to know in
+  which states copy groups will have access to the item.
+  [gbastien]
 
 4.1.28.18 (2021-02-12)
 ----------------------